Enable wakeup from soft off via keypress
Adds the soft_off_wakers node to the device tree, setting the keyboard matrix (kscan0) as a wakeup source.

This allows the keyboard to be powered back on by pressing any key after it has been put into the soft-off state. This configuration is required to make the CONFIG_ZMK_PM_SOFT_OFF feature usable without having to physically reset or replug the device.